Add underline to bar
Through a simple hack (thanks Forell) it is now possible to use an
underline for the whole bar. Height of the underline should usually be
an even number so the text can be offset appropriately, but if it is not
it will just look wonky, not crash.
Configuration for this underline was added to the config file. The color
used is the "hl_col" and the height of the bar is "underline_height".
Offset is calculated automatically using "underline_height / -2".